+++ .env.example
@@ -1,0 +1,8 @@
+SERVER_ADDRESS=localhost
+SERVER_PORT=8888
+CLIENT_PORT=8889
+
+STT_MODEL_FILE=resources/models/ggml-tiny.en.bin
+
+INPUT_DEVICE_INDEX=1
+OUTPUT_DEVICE_INDEX=1
