+++ README.md
@@ -4,12 +4,24 @@
 
 - [git-lfs](https://git-lfs.com/)
 - [uv](https://docs.astral.sh/uv/getting-started/installation/)
+- [portaudio](https://www.portaudio.com/)
+- [clang](https://clang.llvm.org/)
 
 ## Clone
 
 ```bash
 git clone --recurse-submodules -j8 git://github.com/sjohnson31/hans
 ```
+
+## Run Hans
+
+### Server
+
+uv run --env-file .env --package server hans_server
+
+### Client
+
+uv run --env-file .env --package client hans_client
 
 ## Checkout submodules
