--- client/.env.example
@@ -1,5 +1,0 @@
-INPUT_DEVICE_INDEX=1
-OUTPUT_DEVICE_INDEX=1
-SERVER_ADDRESS=localhost
-SERVER_PORT=8888
-LOCAL_PORT=8888
