--- client/pyproject.toml
@@ -1,11 +1,0 @@
-[project]
-name = "client"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.12"
-dependencies = [
-    "librosa>=0.10.2.post1",
-    "numpy>=2.0.2",
-    "pyaudio>=0.2.14",
-]
