+++ packages/client/pyproject.toml
@@ -1,0 +1,20 @@
+[project]
+name = "client"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = "==3.11.10"
+dependencies = ["librosa>=0.10.2.post1", "numpy>=1.19", "pyaudio>=0.2.14"]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+# These should be globally unique, because all package scripts end up
+# in the same place in the venv
+[project.scripts]
+hans_client = "get_audio:main"
+
+#TODO: This is definitely wrong when building a non-editable wheel
+[tool.hatch.build.targets.wheel]
+packages = ["src"]
