+++ packages/client/setup.sh
