+++ packages/server/pyproject.toml
@@ -1,0 +1,31 @@
+[project]
+name = "server"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = "==3.11.10"
+dependencies = [
+    "durations-nlp>=1.0.1",
+    "number-parser>=0.3.2",
+    "numpy>=1.19.0",
+    "silero-vad>=5.1.2",
+    "soundfile>=0.12.1",
+    "tts>=0.22.0",
+    "whisppy",
+]
+
+[tool.uv.sources]
+whisppy = { workspace = true }
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+# These should be globally unique, because all package scripts end up
+# in the same place in the venv
+[project.scripts]
+hans_server = "main:main"
+
+#TODO: This is definitely wrong when building a non-editable wheel
+[tool.hatch.build.targets.wheel]
+packages = ["src"]
