+++ packages/whisppy/src/whisppy/_whisppy.pyi
