+++ pyproject.toml
@@ -1,0 +1,10 @@
+[project]
+name = "hans"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = "==3.11.10"
+dependencies = []
+
+[tool.uv.workspace]
+members = ["packages/*"]
