--- server/.env.example
@@ -1,2 +1,0 @@
-SERVER_PORT=8888
-RETURN_CHANNEL_PORT=8888
