--- server/pyproject.toml
@@ -1,15 +1,0 @@
-[project]
-name = "server"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = "==3.11.10"
-dependencies = [
-    "durations-nlp>=1.0.1",
-    "number-parser>=0.3.2",
-    "numpy>=1.19.0",
-    "pywhispercpp>=1.3.0",
-    "silero-vad>=5.1.2",
-    "soundfile>=0.12.1",
-    "tts>=0.22.0",
-]
